transfer code from tagCloudClient to Actions
