+++ TagsCloudContainer/Client/GUITagCloudClient.cs
@@ -24,7 +24,7 @@ namespace TagsCloudContainer.Actions
             this.wordProcessor = wordProcessor;
         }
 
-        public void DrawImage(string sourceFilePath, string boringFilePath, int imgWidth, int imgHeight)
+        public void DrawImage(string sourceFilePath, string boringFilePath)
         {
             var wordsCount = wordProcessor.CalculateFrequencyInterestingWords(sourceFilePath, boringFilePath);
             var rectangles = cloudLayouter.GetRectangles(wordsCount);
@@ -33,62 +33,7 @@ namespace TagsCloudContainer.Actions
 
         public void SaveImage(string filePath)
         {
-            var dialog = new SaveFileDialog
-            {
-                CheckFileExists = false,
-                InitialDirectory = Path.GetFullPath(filePath),
-                DefaultExt = "png",
-                FileName = "image.png",
-                Filter = "Изображения (*.png)|*.png|Изображения (*.jpg)|*.jpg|Изображения (*.bmp)|*.bmp"
-            };
-            var res = dialog.ShowDialog();
-            if (res == DialogResult.OK)
-                pictureBox.SaveImage(dialog.FileName);
+            pictureBox.SaveImage(filePath);
         }
-
-        public string SetBoringFilePath(string filePath)
-        {
-            var dialog = new OpenFileDialog()
-            {
-                CheckFileExists = true,
-                InitialDirectory = Path.GetFullPath(filePath),
-                DefaultExt = "txt",
-                FileName = "boring.txt",
-                Filter = "Текстовые файлы (*.txt)|*.txt"
-            };
-
-            var res = dialog.ShowDialog();
-
-            if (res == DialogResult.OK)
-                return dialog.FileName;
-
-            return "";
-        }
-
-        public string SetSourceFilePath(string filePath)
-        {
-            var dialog = new OpenFileDialog()
-            {
-                CheckFileExists = true,
-                InitialDirectory = Path.GetFullPath(filePath),
-                DefaultExt = "txt",
-                FileName = "source.txt",
-                Filter = "Текстовые файлы (*.txt)|*.txt"
-            };
-
-            var res = dialog.ShowDialog();
-
-            if (res == DialogResult.OK)
-                return dialog.FileName;
-
-            return "";
-        }
-
-        public void SetSettings<TSettings>(TSettings property)
-        {
-            SettingsForm.For(property).ShowDialog();
-            pictureBox.RecreateImage(imageSettings);
-        }
-
     }
 }
